[HotFix] 스웨거 파라미터 데이터타입 수정

+++ src/test/java/com/umust/dobonglife/domain/business/controller/BusinessControllerTest.java
@@ -34,6 +34,8 @@ import static org.mockito.BDDMockito.willDoNothing;
 import static org.springframework.restdocs.operation.preprocess.Preprocessors.*;
 import static org.springframework.restdocs.payload.PayloadDocumentation.*;
 import static org.springframework.restdocs.request.RequestDocumentation.*;
+import static com.epages.restdocs.apispec.ResourceDocumentation.parameterWithName;
+import static com.epages.restdocs.apispec.SimpleType.INTEGER;
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.multipart;
 import static org.springframework.test.web.servlet.result.MockMvcResultHandlers.print;
@@ -339,8 +341,8 @@ public class BusinessControllerTest {
                                         .summary("사업장 프로모션 목록 조회")
                                         .description("사업장의 프로모션 목록을 조회합니다.")
                                         .queryParameters(
-                                                parameterWithName("lastId").optional().description("커서 - 마지막 프로모션 ID (첫 요청 시 생략)"),
-                                                parameterWithName("size").optional().description("조회 개수 (기본값: 3)")
+                                                parameterWithName("lastId").optional().description("커서 - 마지막 프로모션 ID (첫 요청 시 생략)").type(INTEGER),
+                                                parameterWithName("size").optional().description("조회 개수 (기본값: 3)").type(INTEGER)
                                         )
                                         .responseFields(
                                                 fieldWithPath("success").type(JsonFieldType.BOOLEAN).description("요청 성공 여부"),
