+++ src/test/java/com/umust/dobonglife/domain/coupon/controller/PromotionControllerTest.java
@@ -33,6 +33,8 @@ import static org.mockito.Mockito.when;
 import static org.springframework.restdocs.operation.preprocess.Preprocessors.*;
 import static org.springframework.restdocs.payload.PayloadDocumentation.*;
 import static org.springframework.restdocs.request.RequestDocumentation.*;
+import static com.epages.restdocs.apispec.ResourceDocumentation.parameterWithName;
+import static com.epages.restdocs.apispec.SimpleType.INTEGER;
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.post;
 import static org.springframework.test.web.servlet.result.MockMvcResultHandlers.print;
@@ -105,9 +107,9 @@ class PromotionControllerTest {
                                     .description("프로모션 목록을 조회합니다.")
                                     .queryParameters(
                                             parameterWithName("lastId").optional()
-                                                    .description("커서 - 마지막 프로모션 ID (첫 요청 시 생략)"),
+                                                    .description("커서 - 마지막 프로모션 ID (첫 요청 시 생략)").type(INTEGER),
                                             parameterWithName("size").optional()
-                                                    .description("조회 개수 (기본값: 2)")
+                                                    .description("조회 개수 (기본값: 2)").type(INTEGER)
                                     )
                                     .responseFields(
                                             fieldWithPath("success").type(JsonFieldType.BOOLEAN).description("요청 성공 여부"),
@@ -198,7 +200,7 @@ class PromotionControllerTest {
                                     .summary("쿠폰 발급(포인트 교환)")
                                     .description("포인트를 사용하여 쿠폰을 발급합니다.")
                                     .pathParameters(
-                                            parameterWithName("promotionId").description("프로모션 ID")
+                                            parameterWithName("promotionId").description("프로모션 ID").type(INTEGER)
                                     )
                                     .responseFields(
                                             fieldWithPath("success").type(JsonFieldType.BOOLEAN).description("요청 성공 여부"),
